Blending two images of different size
This file contains the code for blending two images of different sizes. In this code the basic operations  like resizing, cropping, color conversion and thresholding of opencv are used. Both of the images are also uploaded in the resources folding by the name of 'dino.webp' and 'house.jpg'.
Thanks for reading and accessing...