Add files via upload
This is a small project of demonstration of Probabilistic Houghline transform.
First method mask everything other than the region of interest, It get the same colors 
as that of the image provided by in the region of interest.
second method creates a blank image of the size of the original image and draw green lines
at their respective positions which are detected in the original image. At the end we add this 
blank image with lines to the original image.
Third method defined the vertices for the area of interest where we want to draw the 
lines of the road lanes. secondly it converts the image to gray scal so that we can appy the 
Canny edge detector. lastly the probabilistic Hough lines transform is applied to detect
the lines which fullfill some criteria.